use a config module folder
